Basic Auth flow with Okta.
Okta settings need to be set in the appsettings.json.

It doesn't actually do the login step right now (setting the cookies), but it does go through the entire OIDC flow to login the user and retieve the id token.  I still need to do more work on how to actually sign-in the user.  I know how to do it, but I haven't done the work yet to deserialize the JWT token and set the claims.